ShardedTensor Integration with distributed
Summary:
This diff integrates with ShardedTensor in PyTorch distributed according to the plan/discussions in https://fb.quip.com/fwucARGO5SeO

We are doing the first part of integration, which we replaces the ShardedTensor/Metadata defintions in torchrec and using the ones defined in Pytorch distributed.

A second part integration might be more involved that we need to accomodate fbgemm kernels to take sharded tensor and do the computation, then switching to a mode that a ShardedModule contains a sharded weight/tensor directly, instead of multiple small nn.EmbeddingBags.

Reviewed By: YazhiGao

Differential Revision: D29403713

fbshipit-source-id: 279643bd01261ae564238b9dea9d2af5597342c2