Replace shard metadata placement device with meta
Summary: Planer use a real device type to generate sharding plan and thus shard metadata placement device is the real device. If we use meta device to contruct the model, the placement device could be conflicted with the tensor device. So we need to hack the placement device to meta device in order to pass  the unmatched device verification.

Reviewed By: dstaay-fb

Differential Revision: D31836919

fbshipit-source-id: 68c10fe0f5a75b45fea7107e90c05ef5bc58b6cf