Simpler way to check if installation is successful
Summary:
Currently verifying if installation is successful involves installing stl tasks and lightning and is fairly complicated. Apart from the complexity we shouldn't need a second library to verify if TorchRec is installed.
I see a lot of value in having a simpler script to verify if installation is successful

Reviewed By: rkindi

Differential Revision: D32122481

fbshipit-source-id: b4431433be5a799840d6cf6d4c1a5570f8c15afd