Add train/val/test splits for dlrm_main.py (incl. val/test metrics).
Summary:
Add train/val/test splits to dlrm_main.py. As part of adding these splits, I also added val/test AUC + Accuracy TorchMetrics.

These changes are needed so we can evaluate AUROC/Accuracy of the model over the validation set as it trains.

Note: In the metrics, we hold the accumulated preds/labels in GPU memory for the entire val or train set. This is fine because Criteo has ~4.4B rows each ~40 bytes. (4.4B rows / (24 files)) * (1/2 because we take half of the last file for either test or val) * (8 bytes for label + pred) = 733 MB across all ranks (usually a minimum of 8 ). And note this is when the model is in `.eval()` so we don't have the storage requirements of backwards pass so this amount of overhead is reasonable.

Reviewed By: colin2328

Differential Revision: D32661638

fbshipit-source-id: 9222ed3c3c9d1880fb0139c05706655f1a2947bd