Add support for callbacks in Awaitable
Summary:
We extend the Awaitable interface to contain a list of callbacks that we can chain at the end of a wait. The use case for this is to make the interface of quantization better as well as supporting CW sharding + EBC (see later diffs).

I couldn't find a way to keep wait() as the abstractmethod (and use some kind of destructor way of running the callbacks at the end). Thus I changed the abstract wait method to _wait_impl and changed all of them of any sub classes

----

Changed the implementation of quantized_comms to access the property method rather than directly modifying the private instance

Modified every instance of def wait of instance classes to be _wait_impl.

Note that this change in theory is backwards compatiable. If an interface does not change _wait_impl (but rather changes wait), we still see the desired behavior, just without any callback logic (which doesn't exist today anyway)

Reviewed By: xing-liu

Differential Revision: D32960556

fbshipit-source-id: 318142035e8861ca896e667280c8c7a289c5e2e2