improve IFR inference perf
Summary:
we see ~0.5% predictor CPU regression in IFR QE with the latest model.  We could also re-pro the inference perf gap with `ptvsc2_predictor_bench` tool offline.

By moving the kjt.getitem to C++ resolved most of the regression. Here is the offline benchmark runs of hpc/trec-before/trec-after.  After this change, **the inference perf gap reduced from 8.3% to 1.0%**

For long term, we may want to move the whole KJT to c++, which could also let us to enable static-runtime. I had a draft diff D32294637, but it regressed the training QPS, which need some time to investigate.

{F681881861}

Reviewed By: divchenko

Differential Revision: D32582665

fbshipit-source-id: a789a7c3bb821242ce209bc208747b1488136d17