Fix for RW rows less than world size
Summary:
Bug in planner, where shards isn't being populated.

Example case is if we have a table of size 2x64 with num_ranks = 4

We expect shards to produce length 4, with the two last entries indicating empty shards, but current we only produce a shard length of length 2.

This leads to an indexing error in rw_sharding.

Reviewed By: dstaay-fb

Differential Revision: D33098867

fbshipit-source-id: c90375697cef693a158e7f6c88453187d856eb1d