Update torchrec's submodule fbgemm commit id
Summary:
**context**
fbgemm was failing to install out of torchrec's setup.py command because the THC.h header file was being included in the fbgemm repo but it was removed: https://www.internalfb.com/intern/qa/291103/thcthch-not-found----fblearner-flow-compilations-b

This was fixed here: D32872476 but was not reflected in torchrec's checkout of fbgemm as that has to be manually updated.

**this diff:**
This updates the commit id to be the latest.

I'll take a look at how to automate this process if possible. The installation process can break if our version of fbgemm gets too stale.

Reviewed By: jianyuh

Differential Revision: D33251641

fbshipit-source-id: 2c20779ef69071160870275d54bff20cfb4acbc1