Fix TWRW local size bug [RFC]
Summary:
Right now it's a bit hard to test TWRW because we don't have access to multiple nodes to test on.

It's hard for us to replicate the "local_size is 2" case on a single gpu without creating some sub groups on the gpu itself.

Though right now the main issue is some tensor/metadata device placement checks. (Even though we're overriding local_world_size, the true local_world_size is still 8, and I imagine it's hard to get around that unless we mess around with GPU device configs)

So here we just override what planner says the tensor metadata device placement is with the true device that it will be placed on.

Question -

Why are there no tests for gloo_twrw? I suppose it's something that will never be used for GLOO, but could still help us unit test collective call logics.

Reviewed By: colin2328

Differential Revision: D33228397

fbshipit-source-id: 6a420298099120ceba96cb09324d36c1727e7e0e