Incorporate _embedding_permute_op in CW to support EBC + CW
Summary:
Currently in EBC, CW is not fully working. Root cause of the problem is that KT assumes that a feature name appears exactly once in its keys
More concretely, consider that we have one table (call it TABLE_1) of dim 10 x 32, with a world_size of 4:
Since we treat each shard as a separate table, after the output dist we have
Rank 0:
keyed_tensors = [vec_0, vec_1, vec_2, vec_3]
emb_names = [TABLE_1, TABLE_1, TABLE_1, TABLE_1]
However, after calling accessing TABLE_1 through KT (whether through regroup, or getitem or todict), we only keep
vec_0 resulting in the loss of our dimensions.

PEA currently doesn't have this issue as it does the column permutation here
https://www.internalfb.com/code/fbsource/[3ad2e7cc8ba2]/fbcode/torchrec/fb/distributed/pooled_embedding_arch.py?lines=322

However, since this is CW specific logic, we should move it into CW sharding, which is what we do here.

This does not effect PEA's implementation, and after which this change is applied, PEA's output_dist becomes a no-op in the case of CW. If there is mixed sharding schemes, the permute will first be applied to the CW tables and joined. There will then be a second round of permutting to ensure global feature sequencing

-----

*follow ups*
1. Remove merging shards together in planner (I think this breaks something in global tensor init checks)
2. Ensure that training quantization works and adheres to this new callback interface
3. Add logic to planner to not make shards on batched kernels with dim < 4

Reviewed By: YazhiGao

Differential Revision: D32817443

fbshipit-source-id: c8d0fa0beefa26330db49741ccc0b6fd8293b45f