Add support for key -> multiple shards in ShardedModuleWrapper and FusedOptimizer
Summary:
Motivated by the requirement of D32968927 to have multiple shards of the same table per rank, we add an idx to the local shards of ShardedTensor when converting to LocalShardInfo.

Embedding Lookup Optimizer:
We convert emb_weights into a ShardedTensor and map that to the corresponding ShardedTensor representation of the ROW_WISE_ADAGRAD momentum. As a followup, we should make it so that embedding_weights_split can return a ShardedTensor [?].

In TrainModuleWrapper, we currently only export one local shard per ShardedTensor. To support this use case, we add an idx to the key like we do in TrainOptimizerWrapper.

In the future we want to migrate off of LocalShardInfo and onto ShardedTensor completely (there are of drafts of this working). However, there is currently a large overhead of downstream dependencies that are hardcoded to expect LocalShardInfo out. This will be a temporarys solution to avoid that blocker.

A couple questions of concern:
1. There are some downstream test utils that use sharded_state_dict from TrainModuleWrapper and directly use it in load_state_dict for an inference module. Since the names now contain the sharded tensor index, this no longer works. Is this a valid use case? Shouldn't all infer model be loading their state from checkpointing? (this works)
2. Any issues with having self.state EmbeddingFusedOptimizer having ShardedTensor as the key instead of a tensor?

Reviewed By: dstaay-fb

Differential Revision: D33285503

fbshipit-source-id: 68e108244fad9041bfa85fd34a313dff3571a8ff