Don't override tensor properties with local_shards dtype for optimizer dtype
Summary:
Optimizer dtype should not be tied to the dtype of the embedding table shards. However, now because weights is a sharded tensor, we potentially override it's dtype with the local_shard dtype.

Since we use table's tensor property for the optimizer, we inadvetantly override it as well. Here we set it so that the tensor property used for weight is a copy, to not affect that of the optimizer dtype.

However, a couple Qs:

1. We're assuming dtype of optimizer is determined by the embedding table, why? How is this configured?
2. It seems like that table_config.global_metadata can actually be inconsistent with that of local_metadata (in terms of dtype), which is strange.

Reviewed By: dstaay-fb

Differential Revision: D33473885

fbshipit-source-id: 20aeec85fcd44ab0d41931ce46dc4e892d677ad7