Move notation from model to __init__.py
Summary:
In the public docs, I thought that the model class docs would show up on a different page than the __init__.py docs for the models directory. It doesn't, they show up in the same page, so we don't need to repeat the notation information.

Also the notation shows up all in one line:

{F703712065}

bullet points would be nicer.

Differential Revision: D34390595

fbshipit-source-id: 1f1c439bb23aaa27abc6b40434539c10c1cf2361