Fix forward torchrec tracer issue with proxying buffers (#139)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/139

Disable tracing buffers (introduced in https://github.com/pytorch/pytorch/pull/73612) in torchrec train_pipeline.Tracer

This was causing failures in the following test: https://www.internalfb.com/intern/testinfra/diagnostics/5066549672053905.562950011720831.1646546163/

Note that this is a stop-gap solution. In general, buffers should probably be traced. The issue is introduced because the model being traced is mutating an attribute of the module, which is generally unsafe within a functional tracing context:

https://www.internalfb.com/code/fbsource/[830447276a410da5c1a796647782a639a6968054]/fbcode/hpc/torchrec/models/ads/ads_10x_2021h1_jun_v0.py?lines=6626

Reviewed By: colin2328

Differential Revision: D34741803

fbshipit-source-id: f6bf8f6d2689344e0237d6f7cadbc918883e3d12