DLRM return prediciton values
Summary:
Return the prediction values, not just the logits from the pipeline.

From the docs of AUROC and Accuracy:

```
Forward accepts

    - ``preds`` (float tensor): ``(N, ...)`` (binary) or ``(N, C, ...)`` (multiclass) tensor
      with probabilities, where C is the number of classes.
```

In reality it's expecting the probabilities (where the boundary is 0.5). Right now we're returning logits (where boundary for positive is 0). Maybe there are some cases where DLRMTrain model is outputting logits in the region (0, 0.5), which we tag as negative label, but it's actually positive.

Reviewed By: colin2328

Differential Revision: D34968264

fbshipit-source-id: b4375e44062f4451cf232ff8578b767c80ba6702