Convert input/output in stats to be MB per iteration
Summary:
Instead of naively showing input as just the pooling factor and output as just the embedding dimension
this diff changes planner stats to use the actual size of input & output in terms of megabytes per iteration

**input**: global_batch_size * pooling factor * sizeof(dtype of input)

**output**: global_batch_size * (output size (1 in pooled)) * sizeof(dytpe of emb) * emb_dim

This provides a sense of scale for data coming in and out, and additionally makes plans with multiple sharding types directly comparable.

Also fixes a bug with TWCW, we incorrectly specified the ranks as entire world size when it should be limited to the local world of the  host that the parameter is sharded on.

Differential Revision: D35153224

fbshipit-source-id: 531626ef9d1a6ddf8e9feb756f778d9aa2e406a5