Allow random dataloader to have ids per feature to vary between features
Summary:
Currently the random dataloader is inflexible in ids_per_feature because every table will have the same pooling factor.
Adding a bit of a refactor to change that.

Also I found the generation script a bit hard to follow, so changed it - but not sure if this is more inefficient
As a follow up, we can allow a distribution that we draw from to generate ids_per_feature (instead of fixed)

Reviewed By: bigning

Differential Revision: D35148294

fbshipit-source-id: 546c23f75fbad156d27648a63b7d11e85d7ff7c6