Make keyed_jagged_tensor more friendly for torch.jit.script (#196)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/196

KeyedJaggedTensor is not working properly with torch.jit.trace. The reason for this - there is a number of operations that are getting recorded by jit.trace and won't be recomputed the next time. As a result KeyedJaggedTensor is not working well on data with different batch sizes (since stride is not getting computed properly).

This diff is replacing stride computation to a scripted function in case of torch.jit.trace, so it won't be memoized and will be forced to recompute every time.

Reviewed By: kflu

Differential Revision: D35414790

fbshipit-source-id: b4ef337c902765881783b482f8009cc418d74684