Do not wait for H2D copy before calling inference on model
Summary:
Currently wait for H2D to finish before calling inference with the batched inputs. This feels fine but not ideal as there're many cpu side of work when we calling inference (like launching kernels).

We can use event to synchronize between the stream and doesn't have to wait until H2D is done.

Differential Revision: D34627934

fbshipit-source-id: 40fd0f6b5f42f02d05b3f0fbbd391e111458e4d3