Do not wait for H2D copy before calling inference on model (#205)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/205

Currently wait for H2D to finish before calling inference with the batched inputs. This feels fine but not ideal as there're many cpu side of work when we calling inference (like launching kernels).

We can use event to synchronize between the stream and doesn't have to wait until H2D is done.

Reviewed By: YazhiGao

Differential Revision: D34627934

fbshipit-source-id: e44b0b6dd278fa847ae48953d2dabbcc4d619e26