Optimize distributed embedding module for large embeddings (#216)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/216

At the moment TorchRec is using Slice operator to split results of all-to-all communication into JaggedTensors. The problem with this approach that Slice backward operator is operating on dense gradients, which can result in a pretty slow backward pass if there is a large all-to-all operator and large number of embeddings involved.

Replacing this operator to a single Slice operator improves efficiency of the test job by ~2x, compared to the existing implementation.

Reviewed By: RenfeiChen-FB

Differential Revision: D35570058

fbshipit-source-id: e2c2152c855cc57dbc64fdbe802b2f686df43427