run unit tests on python 3.7
Summary:
we sometimes fail binary build (https://github.com/pytorch/torchrec/runs/6288714900?check_suite_focus=true) because of unit tests failure for python 3.7. however, we don't run unit tests on 3.7, so we fail to catch on diff land.

This moves tests to run on 3.7. Long term, we should run on 3.7/3.8/3.9 (cc YLGH we should refactor GHA to do the matrix)

Reviewed By: fegin, zyan0, yachyv7

Differential Revision: D36138351

fbshipit-source-id: 43d39d0f24f24b4f996705efe49a2defcdb22d92