calculating and adding planner runtime to planner summary (#338)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/338

created a script that calculates a runtime of a planner. this is done with "perf_counter" function from "time" library. the calculated runtime is then passed to "stats.py" to output time in auto planner summary.

to calculate runtime start time is defined and initialized in _init definition (here precision is lost by little amount). end time is defined and initialized after we know that best plan is found and just before stats are transferred to "stats.py".

runtime is given in seconds with 2 decimal points precision.

alternatives:
- using "timeit/default_timer" function from "timeit" instead of "perf_counter" function from "time" library
- calculating and wrapping start and end time just around the EmbeddingShardingPlanner

why the committed approach is chosen over the above-mentioned alternatives:
- "perf_counter" is shown to have better precision than both "timeit" and "default_timer" functions. ("default_timer" uses "perf_counter" as its base)
- even though precision is lost, sending runtime to "stats.py" is simpler this way and lost precision is negligible

Differential Revision: D36392606

fbshipit-source-id: 033e8ee5f64f96fb8a0d86c908f41f6f696960e2