EBC based off of SplitTable fbgemm kernel (#367)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/367

Two classes created:

BatchedFusedEmbeddingBag: This is the equivalent of the wrapper in the distributed code base around SplitTableBatched. Ideally we'd be able to share the implementations between the two, but at the moment, the other one has too much specific ShardedTable logic

FusedEmbeddingBagCollection: Utilizing the above, we group tables by their keys (like in the distributed case) and do all batched look ups.

Follow ups:

1. Support more than just SGD Optimizer
2. Implement the unsharded equivalent of FusedEmbeddingOptimizer

One thing we may need to support is for Placement to be meta (to be used when device is also set to meta)

Reviewed By: divchenko

Differential Revision: D36396172

fbshipit-source-id: c365e947659eabe6798af92a32aa601e29fea2ef