Add the should_validate_update flag for states update input checking (#408)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/408

In `rec_metric.py`'s `_update(...)` function, the following 2 if-statements let the has_valid_weights tensor to be passed from GPU to CPU, which leads to CPU-GPU synchronization. This can cause performance reduction. So in this diff, instead of open this feature to all models, we gate the update input checking feature with the `should_validate_update` flag
- `if torch.any(has_valid_weights)`
- `if has_valid_weights[0]`

Reviewed By: fegin

Differential Revision: D36992375

fbshipit-source-id: 4a87733304333991f1d106e2b4d54f8f06f83974