add parallelized EmbeddingShardingPlanner class (#391)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/391

Currently, planner is running in a single threaded process. To improve the planner's runtime and scalability, this diff creates a new class that similar to EmbeddingShardingPlanner with the addition of multi-processing.

This is done by first making a proposals' list from a list of proposers, and then grouping the list into a user's custom cpu number of lists, or into an automatically calculated cpu count with the help of os library.

Grouped prosolals are distributed to cpus and are simultaneously worked on. From each group best plan is chose with help of "get_best_plan" function.

Finally, the absolutely best plan is chosen among best_plans given by groups.

(For more clarification, partitioner is changed to adjust to multi-processing. As parallelized cpus cannot work on shared variables, function, and more, program was outputting wrong calculations. Partitioner is changed so that functions and variables are not shared among processors anymore)

Reviewed By: YLGH

Differential Revision: D36841610

fbshipit-source-id: 3d9f243cf0ec67357ff2317f84684e7566ad2b67