Criteo preprocessing improvements (#402)
Summary:
I'm using the fact how Criteo dataset looks like to improve preprocessing:
* there is no need to cast sparse features to `str` first  as `csv.reader` [here](https://github.com/pytorch/torchrec/blob/50ff0973d5d01ec80fe36ba5f1d524c92c799836/torchrec/datasets/utils.py#L216) produces a list strings
* we can replace missing values -- empty strings -- to zeros right away and then cast features to int avoiding try/catch block in `safe_cast` helper function.

This results in a decent 15% speedup according to my measurements. It is a non-negligible time saving as the preprocessing is a long job. Moreover, I think that this is slightly more transparent in how .npy files are produced.

Pull Request resolved: https://github.com/pytorch/torchrec/pull/402

Reviewed By: joshuadeng

Differential Revision: D37109621

Pulled By: YLGH

fbshipit-source-id: 68f4aafab68161594ff4fecf317742047a3762ef