Add ShardedFusedEmbeddingBagCollection and FusedEmbeddingBagCollectionSharder (#413)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/413

This does the logic of swapping out FusedEmbeddingBagCollection w/ ShardedFusedEmbeddingBagCollection.
For non-batched_fused kernels, we register a backward hook, either manually, or via _register_fused_optim from DDP.
There is currently a gap in the ability to checkpoint the optims from DDP, but this is a WIP

Implementation decisions:

Currently we do all of this in FusedShardedEmbeddingBagCollection. If we REALLY wanted to make things as modular as possible, we would implement BatchedFusedDenseGroupedLookup, and use that as a lookup instead of BatchedDens. But that might be a bit of overkill :p

We choose to manually wrap the data_parallel tables with DistributedDataParallel, and add these tables to sharded_parameter_names, which are then ignored in the DefaultDDPWrapper

Reviewed By: divchenko

Differential Revision: D36482467

fbshipit-source-id: a19188bfe6d89afaaa07c8176c941e20fb47b7e2