change partitioner class methods to function
Summary:
This diff changes unnecessary GreedyPerfPartitioner class methods to function - gets rid of GreedyPerfPartitioner, and only leaves needed functions. Subsequently, GreedyPerfPartitioner was changed to a partition function in parallelized planner and planner programs.

For testing purposes, partition() returns a tuple of sharding plan and topology. both topology and partition are then used in test_partition file to test the program.

Differential Revision: D37165570

fbshipit-source-id: f0fc85ec4b7d377f0e1964a07cfd4ffd9f4af171