Add column wise sharding support for EmbeddingCollection (sequence embeddings) (#432)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/432

Support for CW and TWCW sharding in EmbeddingCollection.

Added logic to stitch feature outputs with local embedding dim after output dist to match original dim.

Outputs are stored in order of rank, however, in column-wise sharding, there can be multiple shards of a table on the same rank and thereby multiple outputs on the same rank.

i.e.
        rank 0: [f_0, f_0, f_1]
        rank 1: [f_0, f_1]
        output: [f_0(shard_0), f_0(shard_2), f_1(shard_0), f_0(shard_1), f_1(shard_1)]

        f_0 shard ranks = [0, 1, 0]  # [f_0(shard_0) = rank0, f_0(shard_1) = rank0, f_0(shard_2) = rank1]

        f_0 output ranks = [0, 0, 1]  # [f_0(shard_0) = rank0, f_0(shard_2) = rank0, f_0(shard_1) = rank0]

        # To get the correct order of outputs we want permute indices for output_ranks -> shard_ranks
        permute_indices = [0, 2, 1]

Since outputs are stored by rank, the inter-shard order is lost and the shards on rank 0 would be combined first, making an incorrect combination of f_0's output with the shard ranks = [0, 0, 1].

To keep the correct shard rank of [0, 1, 0] when combining outputs, we generate permute indices for each feature to match the shard ranks.

Reviewed By: dstaay-fb

Differential Revision: D36944684

fbshipit-source-id: f5af89e1fe01dc9ee2e9a72ed43adaf4a06d24e0