Implement adaptive compute interval (#531)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/531

The compute interval is currently decided by compute_interval_steps. This is
not good as the model may be generated once and used in different scenarios
(e.g., offline, online training). Different scnearios may have different
training speed and thus a fixed compute_interval_steps is not ideal.

In this PR, we propose an adaptive way to change the compute interval. The
idea is as follow:
1. RecMetrics will record the time interval between the first two computes.
2. If the interval is not within the range of min_time and max_time, RecMetrics
will automatically adjust the report batches count to meet the requirement.

Reviewed By: yachyv7

Differential Revision: D37877504

fbshipit-source-id: 8669b2dbba03670a7d5f8d42efb18fa5176bf1c0