Plumb qcomm_codecs_registry from sharder to comm_modules (#538)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/538

We have a QCommCodecs, which contains
two QuantizedCommCodecIf, for forward and backward.

Throughout our entire codebase, we will access these via a
qcomm_codecs_registry, which has a string (representing comm_op) -> QCommCodecs

We have the default implementation of quantized_comm_codec, based on the FBGEMM translations, and have this constructor as an QCommsConfig. We make the decision to make the QCommsConfig ->  qcomm_codecs_registry outside of the sharder level. However, we provide a very easy interface (get_qcomm_codecs_registery) that maps QCommConfig to this regsitry, so users only directly interact with QCommsConfig, and never CommOpTypes themselves.

The base sharder impl/interface acts are only aware of qcomm_codecs_registry

Also found out that we can't use BF16 as a backward precision on gloo, so just test with FP16.

Reviewed By: colin2328

Differential Revision: D37852363

fbshipit-source-id: 8457a76774f3c6c7704b14caf923c221c54113b8