Add batch size to planner interface
Summary:
Instead of making batch size a part of topology as that is not intuitive, we want to make batch size a part of the planner interface.

In this first diff we add batch size to the planner interface to make it backwards compatible internally, and after the changes have propogated we will remove batch size from topology altogether.

Differential Revision: D38337130

fbshipit-source-id: 938ebdc28fd2f1350da8b0cee564983c827e8735