fix tower input dist
Summary:
before this change, for the shardedEmbeddingTower, in the input_dist, we only do the cross_pg dist. Then in the compute() step, we call the tower.embedding() to do the intra_pg input_dist + embedding lookup. The issue is that, the tower.embedding() will introduce host-device sync and it will be excuted on the compute stream.
In this diff, similar to what embeddingTowerCollection does, we move the intra_pg input_dist to the ShardedEmbeddingTower.input_dist, so that, all the host-device sync introduced by the input_dist will be only in the data_dist stream, not the compute stream.

Differential Revision: D38682390

fbshipit-source-id: e780d996c1d3e6ab55606c9710a01d0c071d6e02