Fix dense storage reservation for nested Sharders (#581)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/581

Currently there is a bug with planner calculating ETC dense reservation by consistently underestimating it. In really bad cases it can think that dense storage reservation could be negative.

Root cause of the issue is that since ETC has nested shardable embedding_modules (both the Tower collection itself, and the EmbeddingBagcollection), the cost is being subtracted off twice in the dense reservation. This is because in most cases DMP is getting both the ETC sharder and EBC sharder.

Here we fix this by walking over named_children in a BFS style fashion and stopping as soon as we hit a shardable module.

Reviewed By: bigning

Differential Revision: D38809036

fbshipit-source-id: 3e96d46b74fe7462a0709ce5098656bb6707887c