update category hash sizes and use num_buckets (#591)
Summary:
From docs of nvt.Categorify:

https://nvidia-merlin.github.io/NVTabular/v0.6.1/api/ops/categorify.html

num_buckets ([int](https://docs.python.org/3/library/functions.html#int), or dictionary:{column: num_hash_buckets}) – Column-wise modulo to apply after hash function. Note that this means that the corresponding value will be the categorical cardinality of the transformed categorical feature. If given as an int, that value will be used as the number of “hash buckets” for every feature. If a dictionary is passed, it will be used to specify explicit mappings from a column name to a number of buckets. In this case, only the columns specified in the keys of num_buckets will be transformed.

max_size ([int](https://docs.python.org/3/library/functions.html#int) or dictionary:{column: max_size_value}, default 0) – This parameter allows you to set the maximum size for an embedding table for each column. For example, if max_size is set to 1000 only the first 999 most frequent values for each column will be be encoded, and the rest will be mapped to a single value (0). To map the rest to a number of buckets, you can set the num_buckets parameter > 1. In that case, topK value will be max_size - num_buckets -1. Setting the max_size param means that freq_threshold should not be given. If the num_buckets parameter is set, it must be smaller than the max_size value.

We should use num_buckets as that matches the original implementation's hashing. With max_size we will have a "frequency" thresholding

Also update hash sizes to that of original mlperf

Pull Request resolved: https://github.com/pytorch/torchrec/pull/591

Reviewed By: samiwilf

Differential Revision: D38946545

Pulled By: YLGH

fbshipit-source-id: fc35c0ba9debdeef66d2cdcb0110812d6a33ca82