Migrate bits op and add infra for c++ code (#633)
Summary:
In this PR, we are migrating the `ctz` and `clz` to the top level of torchrec. They are generic versions of `__builtin_clz` and `__builtin_clz` that work on different sizes of int and compilers. We also added `CMakeLists.txt` and change the folder structure for further C++ code.

To compile and test the current code, one could:

```bash
mkdir build && cd build
cmake -DBUILD_TEST=ON ..
make
ctest -v
```

Thank you for your time on this PR.

cc reyoung colin2328

Pull Request resolved: https://github.com/pytorch/torchrec/pull/633

Reviewed By: zyan0

Differential Revision: D39738418

Pulled By: colin2328

fbshipit-source-id: 286fad2d933cb2a5854507938181d8b89823de01