Support incomplete last batches in criteo datapipe. Align misaligned validation and training set dataloading
Summary:
D39597405 & D39597412 are meant to be used together.

The validation set is the 1st half of the last day's samples. The test set is the second half.   Suppose the dataset were 100 elements. Current code would evaluate rows_per_rank = 100 // 16, and treat rank 8's start index and onward as the second half.  Rank 8's start index would be 48 when it's supposed to be 50.  This diff fixes this issue.

This diff also adds support for including partial last batches.  This means that instead of data_set_size % batch_size being dropped, only data_set_size % world_size is dropped.  A common batch_size is 1024. A common word_size is 8, meaning worst case 7 samples are dropped instead of worst case of 1023 when last partial batch is dropped.

Reviewed By: colin2328

Differential Revision: D39597405

fbshipit-source-id: 18e218aa5953aaed8d7d992f1b7d6782501316fe