Refactor deepcopy logic to improve planning speed (#665)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/665

# Summary
When there are many possible proposals to plan with, e.g. >1000 proposals from GreedyProposer for ads models in [fused, fused_uvm_caching] modes, the planner can run >20 min to come up with a best plan. The speed bottleneck comes from `deepcopy(List[ShardingOption])` in proposers and partitioner. With the diff to reorganize the deepcopy operations, the planning time can be decreased to ~2.5 min.

# Idea
(1) Decrease times of deepcopy
When `List[ShardingOption]` is passed as plan/proposal between proposer, partitioner and planner, two times of `deepcopy(List[ShardingOption])` (one in proposer, the other in partitioner) are done to isolate the state of List[ShardingOption] between planning steps. As deepcopy is expensive, we just need to keep one deepcopy in partitioner. Doing so reduces the planning time from ~20 min to ~10 min.

(2) Custom `__deepcopy__` function in ShardingOption
Even with just one time of deepcopy in partitioner, the overall planning time is still dominated by that. Looking into the partitioner code, we found the only field updated in ShardingOption is `shard.ranks`. To avoid copying big objects (e.g. tensor, module) to save time, a custom `__deepcopy__` function is created in ShardingOption that only deepcopies everything except `tensors` and `modules`, which are passed with the original object. With this custom deepcopy function, the planning time further decreases from ~10 min to 2.5 min.

(3) Freeze `tensor` and `module` in ShardingOption
With idea (2), the `tensor` and `module` fields in ShardingOption are changed to be read-only to avoid modification during planning.

Reviewed By: bigning

Differential Revision: D39822605

fbshipit-source-id: eb99a8efe9ff580a88ebd6fa446059f223337e3a